# src/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Agent API
  version: 0.1.0
servers:
  - url: "{api_base_url}"
    variables:
      api_base_url:
        default: ""
paths:
...
  /messages:
    post:
      summary: Create and enqueue a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessage'
      responses:
        "202": { description: "Accepted" }
        "400": { description: "Bad Request" }
  /messages/{id}:
    get:
      summary: Get a message
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "OK" }
        "404": { description: "Not Found" }
...
  /channels/{id}:
    patch:
      summary: Update a channel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "Updated" }
    delete:
      summary: Delete a channel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "Deleted" }
...
components:
  schemas:
    CreateMessage:
      type: object
      required: [audience_type, content]
      properties:
        audience_type:
          type: string
          enum: [broadcast, personalised]
        destinations:
          type: array
          items:
            type: object
            properties:
              channel: { type: string }
              address: { type: string }
              preferences: { type: object, additionalProperties: true }
        content:
          type: array
          items:
            type: object
            required: [type, body]
            properties:
              type: { type: string, enum: [text, md, json, media] }
              body: {}
        template_ref: { type: string, nullable: true }
        variables: { type: object, additionalProperties: true, nullable: true }
        options:
          type: object
          properties:
            ttl_hours: { type: integer, default: 24 }
            idempotency_key: { type: string }
            confirmation_level: { type: string, enum: [sent, delivered, delivered_or_read] }